Can click and place block, only in 1.7f direction on Z axis, will acount for overlapping
Player can click on a block, which set it to inactive to "remove it". when the player clicks another block, the old block will be placed next to it, in the direction of 1.7f on the Z axis. If that space is already occupied by an existing block, it will repeat the calculation for the position until the new position is cleared.